[fdf] Use FIDL platform device for aml-nna
Make the aml-nna driver use the FIDL version of the platform device
protocol instead of the banjo version. Using PDev banjo depends on banjo
proxying which is not supported in DFv2.

Aml-nna was updated to expose the pdev banjo protocol which it proxies
to the underlying pdev FIDL protocol. This was done to support the
aml-nna's child devices which still use banjo.

Bug: 123400
Bug: 110796
Change-Id: I88f82f2d10f16c7d5e6380fb42bbc41f8f94bd48
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821809
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>