[kernel][vm] Populate slots in interval
Add a helper to populate slots in a sparse interval over a specified
range. After the operation, the interval will be split around the range
to: an interval before the range, an interval after the range, and
individual slots for all offsets within the range. This allows more
efficiently pre-flighting a range in an interval we are going to add
pages to, instead of repeated calls to LookupOrAllocateCheckForInterval
for each offset.

Test: vmpl_unittest
Bug: 122842
Run-All-Tests: true
Change-Id: I50e832e09caf994cb9ee2d4b1b40a8f8f1422865
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/812809
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>