[doc-checker] Handle +log in the git branch URLs
doc-checker detects git URLs to HEAD/docs, since these should be file
paths. In this case, the URL was ..../+log/.... and the code was looking
for either /+/ or /+show/. This makes it look for anything that begins
with '+'.

Change-Id: I14a715e831d0de9a1f12597134bf02c08cb8ff25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/823567
Reviewed-by: Nick Van der Auwermeulen <nickvander@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Fuchsia-Auto-Submit: Clayton Wilkinson <wilkinsonclay@google.com>