[build] new restat command-wrapper
In contrast with output_cacher.py, this implementation:

* Always moves outputs to backup before running the command
  (move preserves mtimes)
* Preserves the original command (no transformations needed).
* Moves backup files to outputs if contents match, otherwise discards
  old backup.

This was motivated by an issue where a tool leaked the output file name
into the output contents, so choosing a different output name would
result in differences.  The new approach avoids that possibility by
running the original command as-is.  In the fast case where an output is
unchanged, there are two file moves (the second one deletes the new
output).  Moves on the same device are fast.

A follow-up change will integrate this back into GN.

Bug: 124006
Change-Id: Ib8e01ecd77c8668f8002af03b1bc1727bb4e5443
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/822777
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Fang <fangism@google.com>