[superproject] Roll llvm-project/libc [libc] Adjust NVPTX startup code
Summary:
The startup code needs to include the environment pointer so we add this
to the arguments. Also we need to ensure that the `crt1.o` file is made
with `-fgpu-rdc` set so we can actually use it without undefined
reference errors.
GitOrigin-RevId: 9353e7a71cd35c7472c2f853d6609cf9ce4dd94a
Original-Revision: be3c04aaf110179bd44eaf8455dd5e7b42289dd5
Roller-URL: https://ci.chromium.org/b/8785867551175106641
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Id270a2753af6e28813dbbad78949ebc358b9131a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/823971