[superproject] Roll scudo [scudo] Early exit from the case can't do page release.
There are heuristics to avoid marking blocks if there's little chance
to release pages. So far, those logics only exist in block-marking
section and we didn't leverage the results of those logics. For example,
in a round of releaseToOS try, we know it's still 128 KB away from the
release threshold. In the next round of releaseToOS, we can early exit
if the number of pushed bytes is smaller than 128 KB without looping
each memory group. This CL adds this heuristic and has reduced amount of
time in checking the status of each memory group.

This CL only applies this heuristic on SizeClassAllocator64.
SizeClassAllocator32 has a smaller region/group size and has little
impact on the default value.

Reviewed By: cferris

Differential Revision: https://reviews.llvm.org/D146312

GitOrigin-RevId: 726a4071de06a9e6c5785c2b35b8f305508cf11e
Original-Revision: efd7aa6b5c8f8c6ac0ab5e3168a61eb7492079bf
Roller-URL: https://ci.chromium.org/b/8785836511735746257
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I5413bf0957c173f9ca42433d38754c8317588ce4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/824107