[perfetto-bridge] Fix Flake in Integration Test
I was able to reproduce the flake consistently by adding a sleep after
perfetto-bridge wakes up, but before registering as trace provider.

This implies that the flake was caused by perfetto bridge being too slow
to start and the test initializing the tracing session before perfetto
bridge could register. Perfetto bridge would finish its registration mid
test, but not be initialized. This causes trace manager to attempt to
start/stop it, but it never responds, leading to the timeouts we see in
the log.

To fix this, this patch:
- Removes the trace provider start up from perfetto_producer since it
  doesn't need to register as trace provider to be a trace observer
- Sets up perfetto bridge integration test to wait until it actually has
  a trace provider before attempting to initialize and start

With this, the test successfully passes, even with a forced sleep before
registering in perfetto bridge.

Bug: 123637
Change-Id: I7555e7c529b4940a39386836cd33a50abc4b439c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/822717
Fuchsia-Auto-Submit: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: Brian Hamrick <bhamrick@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>