[fdf][compat] Add support for calling suspend during prepare_stop hook
The protocol driver_manager uses to determine the reason it's stop hook
was invoked is routed to all drivers. This mechanism necessarily must be
out of band of the component stop hook as part of the component
lifecycle.

Additionally, the Stop hook was modified to allow you to override
behavior for deleting the node. This allows us to avoid calling release
in the suspend path, which is what DFv1 does.

Bug: 97457
Change-Id: Ic2f0c65a0dc5865e54d568f9b27fbae5d1e9dcb4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744579
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>