[devfs] Add devfs-with-pkg component
There are two stages for shutting down drivers, one for boot-drivers,
and another for pkg-drivers. `/bootstrap/devfs` provides the signal for
`/bootstrap/boot-drivers` to be shutdown and the newly added
`/bootstrap/devfs-with-pkg` provides the signal for
`/boostrap/pkg-drivers` to be shutdown. This is necessary because we
cannot add a fake dependency from pkg-drivers to devfs as doing so
would cause a dependency loop when things like fshost try to rely on
devfs. This is why `fshost` still uses `/dev` from devfs after this
change but almost everyone else uses `/dev` from `devfs-with-pkg`.

Bug: 108571
Bug: 108572
Change-Id: I5949d589a9060404860bdcdc14985082b97da6dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/818355
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Dmitry Yatsushkevich <dmitryya@google.com>