[f2fs][virt] Add explicit bash dependency.
We may offer a different shell that doesn't support "set +H" in the
VM based tests. To be compatible with that change we explicitly use bash
for the operations that depend on it.

Once we migrate to another shell (ex: dash), I believe we can drop both
bash and set +H here.

Bug: 123888
Change-Id: I4dcde4e24153a92d36e63ff21ee6b111c24e8179
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/824479
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>