[virtualization] Default to dash on debian
Dash is simpler and smaller. I'm working on a test environment that can
run directly out of the debian initramfs and dash is a simpler solution
there. Fixing this to dash will help keep tests consistent if they're
run against the initramfs or the full debian system.

Change-Id: Ie0b5c9dafba9955d02fb27665322bea4493b2da6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/824525
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>