[fx][serve-updates] Fix checking for running ffx server
This updates the codepath where we are starting a `pm` server but are
configured to use the `ffx` server to stop checking the ffx config to
see the configured port. This is now unnecessary, since we can directly
ask the server which port we are supposed to be running on.

Change-Id: Ib9087e007cd4d3324bb602afdc4999155d2faf64
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825363
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Ben Keller <galbanum@google.com>