[fdf] Use sysmem FIDL in amlogic-video
Make the amlogic-video driver use the FIDL version of the sysmem
protocol instead of the banjo version. Amlogic-video depends on banjo
proxying in order to get the sysmem banjo protocol which DFv2 does not
allow.

Change the sysmen fragment in the vim3-video, sherlock-video,
nelson-aml-video, and astro-video composite bind rules to match to the
FIDL version of the sysmem protocol instead of the banjo version. The
amlogic-video driver binds to the composite devices that use those bind
rules and needs access to the FIDL version of the sysmem protocol.

Bug: 123739
Change-Id: I7f545dda20035c237349a3c7cb20cbbc5eb76369
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/824271
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>