[starnix] Improve fidelity of Linux vDSO layout
This tweaks the vDSO linker script in two ways:

Updates the section layout to include the section headers in the
layout and allow the text (and other) sections to follow immediately.
This ensures that the section headers are in the memory layout and also
reduces the size to one page on x86_64.

Also includes the file header explicitly in the PT_LOAD section. This
works around a glibc bug where it assumes that the virtual address
offset of the PT_LOAD program header is always 0. Since this section
includes the file header at the start of the file the linker places
this section at a zero offset.

Change-Id: Ib4038505cdec4a2499221c9d32d2e170e952272f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825356
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Reviewed-by: Maria Glukhova <mariagl@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>