[f2fs] Update 'od' invocation
The busybox od implementation for -i defaults to printing 2b decimal
values, so when dumping 32-bits with that vertion we get the string
'0      0' instead of just '0'.

Using '-td4' instead of '-i' will accomplish the same result but is
explicit in that output should be in 4 byte values.

Bug: 123888
Change-Id: I722e2d057c15c0a32de42d2813c34940751e14b5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825074
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>