[scenic] Improve LoopDispatcherHolder destruction.
Previously, LoopDispatcherHolder would "deadlock" if destroyed on its
own dispatcher's thread.  This is addressed by providing a second
dispatcher which is used to destroy the loop on a different thread.

In addition to the new unit test, this has been manually verified to
fix the leaking of 3 handles every time a Flatland session goes away
(the port, thread, and timer associated with an async loop).

Fixed: 123044
Change-Id: I1b363730905b7e175ad8c2e340c1102634c51687
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/816056
Reviewed-by: Emircan Uysaler <emircan@google.com>
Commit-Queue: Josh Gargus <jjosh@google.com>