[devicetree] Introduce PropertyDecoder
PropertyDecoder is introduced as a layer on top of Properties.
This layer provides an abstraction tied to the Walk itself, such
that in later patches arbitrary properties of nodes along the path
can be used for decoding special properties.

For now only the FindProperties method is introduced.

The Devicetree::Walk Visitor now provides access to a PropertyDecoder,
instead of a Properties. Users can still access the raw properties if
they need to, through PropertyDecoder::properties().

Bug: 78675
Test: fx test devicetree-trests
Change-Id: Ifdccbb1643590fc0448856449f8d11d4fbd432fe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/818970
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>