[fatfs] Fix shutdown flake
I think this flake started showing up when we changed close so that it
just dropes the channel.

To fix this, we need to wait for the scope to finish.  I've also changed
it so that we have only one shutdown path now.

Fixed: 123920
Change-Id: Iabbfead54466a2d12769e7a39ff7f9ce359afadf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/823867
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>