[devshell] Moving test dependency to //tools/devshell
The bash script tests in devshell depended on a
files defined in //scripts/sdk/gn/bash_tests, and a go program
that runs the bash script tests.

The GN SDK bash scripts are being removed, so this moves
the required files to the devshell directory.

Bug: 88715

Change-Id: I8a6e4d5bbff1466383abcb1b5c0d424828f42c57
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825019
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Sébastien Marchand <sebmarchand@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>