[superproject] Roll chromium/deps/icu@default Include BTI flags to assembly file
The generated assembly file is missing proper flags when compiling with
Branch Target Identification enabled.

With this CL compilation flags are added to this file so that the
compiler generates the required GNU poperty section to mark the file
as BTI and PAC compliant.

Original-Bug: 1356083
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromium/deps/icu/+/3852465
Original-Revision: 2958a507f15e475045906d73af39018d5038a93b
GitOrigin-RevId: 13a4a3662fd3aff0f614538122ab8df2c01ea83e
Roller-URL: https://ci.chromium.org/b/8785356050864267729
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Iaf7af726e257058d1ceb51688dc3047fc8dab0b1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/826218