[fidl] Experimental overlay feature
This is an experimental feature only available when the zx_c_types
experiment is enabled. Support is only in fidlc, not in any backends. It
is currently designed and implemented to support parts of the kernel
interface. In the future it may expand beyond that, but let's not get
ahead of ourselves, okay?

Overlays are a union-like structure. The syntax and JSON IR are very
similar to that of unions with a few constraints: no reserved members,
all overlays are strict, overlays may not be optional, no resource
overlays. The wire format is a 64-bit ordinal followed by the specified
member, padded to the size of the largest member.

IFTTT=no changes to good error tests

Change-Id: Ie081949616da318f8bc60b3ece794eb29c670911
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821868
Reviewed-by: Mitchell Kember <mkember@google.com>
Fuchsia-Auto-Submit: Ian McKellar <ianloic@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>