[starnix] Improve prctl(PR_SET_VMA, ..) support
This prctl can be used to assign a name to multiple mappings in one
call and has specific requirements on the input pointer and pointed to
string. The details of this prctl have changed in recent versions. This
updates Starnix to the behavior of a recent version (5.15).

This also propagates the name to snapshots (i.e. forked children).

Test: bionic_unit_tests_static, mm::memory_manager in starnix_tests.cm
Change-Id: Icab055834946bf897c876b000152ce04c981334e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/799043
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>