[perfetto-bridge] Fix Race in Startup
When starting up perfetto, perfetto-bridge uses a condition variable to
wait until perfetto has finished starting up. However, it's possible
that the Posted Task finishes and notifies the condition variable before
the main thread startings waiting on the cv, causing it to miss the
notify and wait forever.

This fixes the issue by adding an additional bool state protected by the
mutex. The state is only read or written while a lock on the mutex is
acquired. If the main thread sees that the state indicates that perfetto
is already initialized, it will instead skip waiting.

Bug: 123637
Change-Id: Iab1d385eb68fbeb1b4857819f69534a4b410680a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825581
Reviewed-by: Brian Hamrick <bhamrick@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>