[assembly] Move sysmem connector into assembly's core_realm AIB
This adds sysmem connector to all systems that have a '/core' realm.

The existing targets for adding sysmem_connector have been moved to
empty groups to enable the transition (and so that they can be removed
as deps in other files in a separate CL).

Bug: 122862
Change-Id: I71a8da59e72433f6b57421a79527e6042590fadc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825182
Reviewed-by: Eric Stone <ecstone@google.com>
Commit-Queue: Aaron Wood <aaronwood@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>