[netstack2] Increase the neighbor event queue size
If the fuchsia.net.neighbor event queue is smaller than the number of
neighbors, than a loss of connectivity will generate `DEL` events for
all neighbors, and flood the queue (causing the Netstack to drop all
clients of fuchsia.net.neighbor).

This CL increase the size of the queue to 4 * the maximum number of
neighbors per interface.

Bug: 121743
Change-Id: I080760f73942fddf051a5f15885a6357a7e18a1c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/824598
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Antonio Linhart <antoniolinhart@google.com>