[starnix] Add a Close method on DevBinder
This will be needed when the protocol will be associated with a Task. At
this point, it will be necessary to correctly order open and close on a
given binder device, so that the same process doesn't end up opening the
device twice because the Open on the DevBinder has been reordered with
the close on the Binder channel.

Change-Id: I5ced78b001d24bb5c504abae9739a11c429639f8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/826021
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
API-Review: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>