[starnix] Fix underflow bug in sendmsg
Due to rounding up the read size to a multiple of sizeof(usize), it can
be the case that the size of "read" data can exceed the total amount of
data. In this case, the remaining size calculation would underflow and
usually lead to an allocation failure.

Fixed: 124385
Change-Id: I1bab7a15704c0be604e526b4a1c0bed1b55e01a7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825079
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Brian Hamrick <bhamrick@google.com>