[component_manager][routing] Simplify RouteSource
Turn RouteSource from an enum of capability types into a struct that
holds the capability source and (possibly empty) relative base path.
This significantly simplifies handoff between the router and its
clients.

As part of this refactor, we need to pull the storage-specific handling
bits out of the router and into component_manager. Helpers have been
added to `storage/mod.rs` for this purpose.

Change-Id: Ibc161228148881b006b4bd04ba3ef62e484ea057
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/820832
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Ali Zhang <alizhang@google.com>
Commit-Queue: Gary Bressler <geb@google.com>