[netstack/bench] Fix throughput calculation
The throughput calculation was erroneously subtracting 'lost_packets'
from 'packets', but the value associated with the 'packets' key in the
iperf3 JSON output means packets that were actually received, and not
total packets.

This incorrect assumption was made due to the fact that iperf3's
'lost_percent' value is calculated as `lost_packets / packets`,
so it would be natural to expect that 'packets' means total packets;
but in reality the 'lost_percent' value is not calculated as one
would normally expect.

Change-Id: Ic1e68fcd598a3dd2e7a6df06c45bd8904bd79cc9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/827079
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Fuchsia-Auto-Submit: Tony Gong <gongt@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Tony Gong <gongt@google.com>