[kernel][vm] Replace supply_zero_offset_
Use sparse zero intervals to represent dirty zero ranges in pager-backed
VMOs instead of a single supply_zero_offset_. This moves all of the page
tracking logic into the VmPageList, which is the source of truth of all
page content, and gets rid of the supply_zero_offset_ which was used to
track content information outside of the VmPageList.

This simplifies a lot of the dirty tracking and writeback logic allowing
the use of unified page traversal loops, instead of treating pages
before supply_zero_offset_ differently from those after.

The awaiting_clean_zero_range_end_ is similarly removed, and
AwaitingClean information is directly tracked in the zero interval. When
zero intervals are written back, they are simply removed from the page
list and user pager will be asked to supply contents if needed.

Test: core-pager-writeback
Bug: 122842
Run-All-Tests: true
Change-Id: Ib7dc5ec93b685473f957cd8e36b6099b2b41cfc6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/803388
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>