[audio] Fix virtualaudio/Control/GetNumDevices
Previously, virtualaudio/Control/GetNumDevices returned the count of
devices that had EVER been created; it did not reduce this number as
device instances closed. This is incorrect behavior. Now the FIDL
API returns the number of currently active (bound) virtualaudio
instances.

Because individual Device instances are unaware of the singleton
Control object, it is not notified when a Device closes. Instead, it
checks whether each device is bound (and cleans it out, if not) in
the GetNumDevices call itself (since this is the only time we
iterate the entire set of devices).

virtualaudio is only used by test code and developer utilities.

Change-Id: I9b57b78c197e9c524e40e352ca7d0f5e71482da2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/827108
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>