[fxfs][blob] Wire Fxfs into assembly
Adds a board `x64-fxfs` which instructs assembly to generate an Fxfs
image rather than an FVM image.  This Fxfs image contains a `blob`
volume which stores all of the blobs required for the system image.

Some changes were necessary to `fx qemu` to use this Fxfs image when
it is generated.

This change also permits Fxfs to use the Vmex resource which is
necessary to make the x64-fxfs board boot.

Bug: 122056
Change-Id: Ic2e3f84ecfe73e16ed04a3a6119741f9f28a7037
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/806092
Reviewed-by: Aaron Drew <ripper@google.com>
Reviewed-by: Dmitry Yatsushkevich <dmitryya@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>