[driver_runtime] Create thread pool and set role profile.
For now, we will create a new thread pool when the user
creates a dispatcher with a scheduler role. We will post
a task to set the role profile on that thread.

I've also updated the existing tests to not pass in erroneous
role profiles.

More work needs to be done to be able to elevate a dispatcher
to another role profile.

Bug: 85946

Change-Id: I4d4da8c36d4428e0a13cc68f2aa23bfa2358d688
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/827025
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>