[starnix] Permit nanosleep with a relative REALTIME offset
The nanosleep syscall is supposed to specify a REALTIME offset and newer
libcs will implement the nanosleep() call with a clock_nanosleep() call
that explicitly specifies this clock. For now, Starnix assumes that the
REALTIME and MONOTONIC clocks advance at the same rate so we can
interpret a relative REALTIME sleep as a relative offset to the
MONOTONIC clock. This updates sys_nanosleep to use the correct clock and
relaxes the sys_clock_nanosleep implementation to permit relative
REALTIME offsets.

Bug: 117507
Change-Id: If2f12fada0736eb89048809a86c7711f91f1b1c4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/827089
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: James Robinson <jamesr@google.com>