[kernel][phys] Convert elf-code-patching-test to physload module test
physload_module_test() gives a convenient framework for loading-related
phys tests: in this change we convert zbi-elf-code-patching-test over to
being specified via the template.

Also addressed in this change was a deficiency in which Symbolize
tracked not the current module, but the initial main one: this made it
difficult for intermediate nodes in a longer load chain to verify that
the expected parent nodes (via PT_INTERP build ID assertions).

Bug: 91187
Change-Id: Ieb5026f5a551e8e49e1dc7647f96c99a4f7ea836
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/795906
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>