[ffx][emu] Sticking the landing of Using instance data for target collection"
This is a reland of fxr/807069 and fxr/814450

The root cause of the issue was a race condition when starting and
stopping emulator instances without using isolate_dir. The target
instances were being matched by name, but the ssh port was different
for the different instances.

The symptom is the emulator starts in under 2 seconds (normal is ~20
seconds), and then ffx target list may show the instance, but
`ffx target show` or other commands that connect to the device
will fail.

Change-Id: Id7e78ab8f0f9eb23444de8d7c6aea069ba42aaab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/818547
Reviewed-by: Steven Grady <slgrady@google.com>
Reviewed-by: Sébastien Marchand <sebmarchand@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>